fix alumni
