+++ app/views/extranet/alumni/persons/show.html.erb
@@ -16,7 +16,7 @@
 
     <div class="experiences mt-4">
       <ul class="list-unstyled">
-        <% person&.cohorts&.each do |cohort| %>
+        <% @person&.cohorts&.each do |cohort| %>
           <li class="experiences__experience py-4 border-top">
             <div class="row">
               <div class="col-md-6">
