+++ test/controllers/extranet/alumni/organizations_controller_test.rb
@@ -8,14 +8,6 @@ class Extranet::Alumni::OrganizationsControllerTest < ActionDispatch::Integratio
     assert_response(:success)
   end
 
-  def test_search
-    get alumni_search_university_organizations_path(term: "Organisation de test")
-    assert_response(:success)
-    results = JSON.parse(response.body)
-    assert_equal(1, results.size)
-    assert_equal("Organisation de test", results.first["label"])
-  end
-
   def test_show
     get alumni_university_organization_path(university_organizations(:default_organization))
     assert_response(:success)
